use int64 as GID

+++ main.go
@@ -10,9 +10,9 @@ type QMsg struct {
 	QmsgVersion 	string            `json:qmsg_version`
 	Type        	string            `json:"type"`
 	Source      	string            `json:"source"`
-	SourceSuccess	bool			  `json:"source_success"`
+	SourceSuccess	bool              `json:"source_success"`
 	SourcePath  	[]string          `json:"source_path"`
-	SourceID    	uint64            `json:"source_id"`
+	SourceID    	int         	  `json:"source_id"`
 	Host        	string            `json:"host"`
 	Msg         	string            `json:"short_message"`
 	Time        	time.Time         `json:"time"`
